+++ src/main/java/com/Django/TraceChain/dto/WalletDto.java
@@ -1,14 +1,15 @@
 package com.Django.TraceChain.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class WalletDto {
     private String address;
-    private long balance;
+    private BigDecimal balance;
     private List<TransactionDto> transactions;
     private List<String> patterns;
 
-    public WalletDto(String address, long balance, List<TransactionDto> transactions, List<String> patterns) {
+    public WalletDto(String address, BigDecimal balance, List<TransactionDto> transactions, List<String> patterns) {
         this.address = address;
         this.balance = balance;
         this.transactions = transactions;
@@ -16,7 +17,7 @@ public class WalletDto {
     }
 
     public String getAddress() { return address; }
-    public long getBalance() { return balance; }
+    public BigDecimal getBalance() { return balance; }
     public List<TransactionDto> getTransactions() { return transactions; }
     public List<String> getPatterns() { return patterns; }
 }
